Updated VocaDbLyricsLib version used.
Also some changes to the build process.
If you look at the build events, there's an automatic ILMerge step done if ILMerge is in your path variable.
It's also set up to copy the plugin to the correct location (assuming MusicBee is installed to "D:\Program Files (x86)\MusicBee"). If you have MusicBee installed elsewhere, you might want to change this, it's just for my convenience.
Doing this let me remove the MusicBee.exe project. Now I have MusicBee set up as the external program to start and due to the copying, nothing needs to be done after a build.